Fix type

+++ lib/DomainDecomposition.py
@@ -767,7 +767,7 @@ def SolveOnCellKeopsGrid(muX,subMuY,subY,posX,posY,rhoX,rhoY,alphaInit,eps,\
     L_logmuX = LazyTensor(torch.log(KemuX).view(4, -1, 1, 1))
     L_posY = LazyTensor(KesubPosY.view(1, 1, -1, dim))
     C_ij = ((L_posX - L_posY) ** 2).sum(-1) / 2
-    eps = torch.Tensor([1/ blur**2]).cuda()
+    eps = torch.Tensor([1/ blur**2]).type_as(KemuX).cuda()
     log_rho = (L_logmuX + L_alpha/eps - C_ij/eps).logsumexp(1) # has shape (4, NY)
     P = KesubRhoY.view(1, -1) * exp(beta.view(1, -1)/eps + log_rho.view(4 ,-1))
     print("Just computing cell marginals")
